pyproject.toml
[project]
name = "Quetzal Analytic Tools"
version = "0.0.1"
description = "django tool to visualize demographics for a specific location"
authors = [
    "Elmer Camargo <elmer@tazacafe.us>"
]
license = "MIT license"
requires-python = ">=3.8"

keywords = ["small-business", "quetzal", "census", "demographics"]

classifiers = [
    "Topic :: Software Development"
]

[tool.poetry]
name = "Quetzal Analytic Tools"
version = "0.0.1"
description = "census wrapper to get demographics for a specific location"
authors = [
    "Elmer Camargo <elmer@tazacafe.us>"
]
packages = [
	{ include = "spatial_tools/census_api.py" },
        { include = "analysis_tools/venmo.py"}
]

[tool.poetry.dependencies]
python = "^3.9"

#GDAL = {path = "dependencies/GDAL-3.4.3-cp39-cp39-win_amd64.whl", develop = true}
#Fiona = {path = "dependencies/Fiona-1.8.21-cp39-cp39-win_amd64.whl", develop = true}

geopandas = "^0.11.1"
census = "^0.8.19"
us = "^2"
python-dotenv = "^0.21.0"
matplotlib = "^3.6.1"
mapclassify = "^2.4.3"
folium = "^0.13.0"
shopifyapi = "^12.1.0"
instagrapi = "^1.16.30"

[tool.poetry.group.dev.dependencies]
spyder = "^5.3.3"
pytest = "^7.1.3"
behave = "^1.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125
